[Ref] Silence warning introduced in previous commit.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19654 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/plugins/MidiInOut.cpp
@@ -429,7 +429,7 @@ bool MidiInOut::MidiSend(uint32 midiCode)
 	}
 
 	mpt::lock_guard<mpt::mutex> lock(m_mutex);
-	m_outQueue.push_back(Message(GetOutputTimestamp(), &midiCode, MIDIEvents::GetEventLength(midiCode)));
+	m_outQueue.push_back(Message(GetOutputTimestamp(), &midiCode, MIDIEvents::GetEventLength(static_cast<uint8>(midiCode))));
 	return true;
 }
